+++ changelog.markdown
@@ -8,6 +8,7 @@
  * tableficate:theme generator now takes an optional `PARTIAL` arg to specify one file
  * `column` now takes a `header_attrs` option to provide attributes for the `th`
  * `actions` now takes options just like `column`
+ * `column` now takes a `cell_attrs` option to provide attributes for every `td` in the column
 
 ## 0.2.1
  * fix table generator to provide `scope` the proper argument and correct documentation